Handle HTTP status code
