+++ src/main/java/com/semmle/jira/addon/LgtmServlet.java
@@ -242,14 +242,4 @@ public class LgtmServlet extends HttpServlet {
     resp.setStatus(code);
     resp.getWriter().write(new Gson().toJson(value));
   }
-
-  public class JsonError {
-    public int code;
-    public String error;
-
-    public JsonError(int code, String error) {
-      this.code = code;
-      this.error = error;
-    }
-  }
 }
